pyproject.toml
[project]
name = "toword"
version = "1.0.0"
description = "PDF/Markdown to Word converter with smart formatting"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.1.0",
    "pdf2docx>=0.5.8",
    "markdown>=3.5.0",
    "pymupdf>=1.23.0",
    "Pillow>=10.0.0",
    "httpx>=0.25.0",
    "mammoth>=1.6.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "latex2mathml>=3.77.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src", "main.py"]
